+++ test/OpenTelemetry.Tests/Trace/CompositeActivityProcessorTests.cs
@@ -99,5 +99,30 @@ namespace OpenTelemetry.Trace.Tests
             Assert.True(p1.ForceFlushCalled);
             Assert.True(p2.ForceFlushCalled);
         }
+
+        [Fact]
+        public void CompositeActivityProcessor_ForwardsParentProvider()
+        {
+            using TracerProvider provider = new TestProvider();
+
+            using var p1 = new TestActivityProcessor(null, null);
+            using var p2 = new TestActivityProcessor(null, null);
+
+            using var processor = new CompositeProcessor<Activity>(new[] { p1, p2 });
+
+            Assert.Null(processor.ParentProvider);
+            Assert.Null(p1.ParentProvider);
+            Assert.Null(p2.ParentProvider);
+
+            processor.SetParentProvider(provider);
+
+            Assert.Equal(provider, processor.ParentProvider);
+            Assert.Equal(provider, p1.ParentProvider);
+            Assert.Equal(provider, p2.ParentProvider);
+        }
+
+        private sealed class TestProvider : TracerProvider
+        {
+        }
     }
 }