+++ src/OpenTelemetry/CHANGELOG.md
@@ -11,6 +11,9 @@
   null-valued tag.
   ([#3325](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3325))
 
+* `CompositeProcessor` will now ensure `ParentProvider` is set on its children
+  ([#3368](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3368))
+
 * Added `ForceFlush` and helper ctors on `OpenTelemetryLoggerProvider`
   ([#3364](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3364))